app/user_profile_service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
gunicorn>=20.1,<21.0
psycopg2>=2.9,<3.0
dj-database-url>=1.0,<2.0
django-environ>=0.10,<1.0
whitenoise>=6.4,<7.0